feat(crm): sync activity moves into a dialog off the connection menu; the mapping card shows a failed round
The activity list is read when something looks wrong, not watched, so it
leaves the page: the connection card's menu opens it in a dialog, which only
queries and polls while open. With the list gone from the page, a failed
full round needs a place to show — the mapping card's status row now takes
the latest full run of its mapping from the run log and reports a failure
with its reason; the mapping itself only records successes. The message log
keeps its shared title (the CRM description says what it holds), since the
run log now carries the sync side by name.
